// src/app/administration/company/model/company-request.ts
export interface CompanyRequest {
  name: string;
  email: string;
  staff: string;
  status: string;
  sector: string;
  responsibleManager: string;
  phone: string;
  address: string;
  city: string;
  zipCode: string;
  website: string;
  legalStatus: string;
  siren: string;
  vatic: string;
  apeCode: string;
  siret: string;
  rcs: string;
  accountNumber: string;
  bic: string;
  rib: string;
  iban: string;
}